Refactor: getRecentTags
최근 태그 목록을 가져오는 API 리팩토링
