+++ server/Tag/tagController.js
@@ -1,15 +1,16 @@
 const { sequelize } = require("../models");
-const GetRecentTags = async () => {
-  const [tags, _] = await sequelize.query("select TagId from PostHashtag order by createdAt desc limit 15;");
-  const contents = [];
-  for (const v of tags) {
-    const [result] = await sequelize.query("select content from Tags where id=?;", {
-      replacements: [v.TagId],
-    });
-    contents.push(result[0].content);
+const tagService = require("./tagService");
+
+const getRecentTags = async (req, res, next) => {
+  try {
+    const recentTags = await tagService.getRecentTags();
+    return res.status(201).json(recentTags);
+  } catch (err) {
+    console.error(err);
+    next(err);
   }
-  return contents;
 };
+
 module.exports = {
-  GetRecentTags,
+  getRecentTags,
 };
